Working on the Dependency Map tool
The goal is analyzing a jOOPL-based library and get a dependency map:

a) What namespaces are in the library across its multiple files?
b) What classes and enumerations are defined in the whole namespaces?
c) And where are the whole classes and enumerations required?

The a) and b) are already answered (i.e. developed), and now's time for
c).

Having a, b and c, jOOPL will be able of determining what JavaScript
files should be dynamically loaded into the Web browser based on what
types are going to be used inside a $namespace.using block!