Documentation changes and code fixes: version 2.4.1
After a big failure with previous exception handling approach, now jOOPL
2.4.1 introduces a very useful and yet powerful exception handling
mechanism.

In the other hand, there were some missing parts of jOOPL in the
documentation:

- The new "Environment" class and its global exception handler.
- More detailed documentation about exception handling!

See the new "Exception handling best practices" section in the sidebar
of official documentation site for more info!

Also, 2.4.1 fixes a hidden issue which was preventing calling the most
derived version of a class member from a class constructor. Now this is
fixed.