Working on 2.5 version
- 2.5 version will feature the beta version of dependency analyzer so
JavaScript resources will be optionally loaded using an state-of-the-art
module definitions and module loading.

- A lot of changes on dependency analyzer. It's being optimized to
maximize its performance. Some fixes have been done as well because it
was not correctly detecting dependencies in order. Now it is capable of
detecting class hierarchy JavaScript files and they can be loaded in the
correct order!