Lib's source code optimizations and GruntJS integration
- Now namespaces and classes are in an associative array like a
hashmap/dictionary so when $namespace.register/$namespace.using requires
to locate some namespaces/class/enum, it won't need to do a search but
just look for something by key. This increases overall performance and
it sounds better, isn't it?

- Some useless code lines have been removed.

- Now jOOPL can be built with GruntJS.

- Node packages have been removed from the Git repository. Use npm
install to install them in your working directory.

- Added a .gitignore file so only changes on desired files and folders
will be included in commits.