Improvements on analyzer
- On some class dependent files, now the file which declare the whole
class appears in the last position. This important so now the file which
declares the class can instantiate that class without troubles!

- Now supports a new configuration parameter called
"baseDirectoryOverrides". This is because sometimes it's required that
base directory of detected files may be replaced by other one (for
example, when deploying to different environments - dev, production...)

- Analyzer test has been updated to reflect the dependency detection
order change.